+++ pkg/k8s/scheme.go
@@ -1,0 +1,15 @@
+package k8s
+
+import (
+	"reflect"
+
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+func ObjectKind(obj runtime.Object) string {
+	t := reflect.TypeOf(obj)
+	if t.Kind() == reflect.Pointer {
+		t = t.Elem()
+	}
+	return t.Name()
+}